Remove unnecessary code & cleanup Action View `abstract_unit`
- Remove `active_record` require --> AV does not depend on AR
- Remove `Rails.env` hack
- Remove translations/localization code, appears to not be used
- Remove `with_autoload_path`, appears to not be used
- Move `Workshop` to the test it's being used.
